// lib/api/auth.ts
import axios from "./axios";
import { LoginData, RegisterData } from "@/app/(auth)/schema";
import {API} from "./endpoints";

export const register = async (registerData: RegisterData)=>{
    try {
        const response = await axios.post(
            API.AUTH.REGISTER, 
            registerData
        );
        return response.data; 
    }catch (err: Error | any){
        throw new Error (
            err.response?.data?.message
            || err.message
            || 'Registration Failed'
        );
    }
}

export const login = async (loginData: LoginData)=>{
    try {
        const response = await axios.post(
            API.AUTH.LOGIN, 
            loginData
        );
        return response.data; 
    }catch (err: Error | any){
        throw new Error (
            err.response?.data?.message
            || err.message
            || 'Login Failed'
        );
    }
}


export const whoAmI = async () => {
  try {
    const response = await axios.get(API.AUTH.WHOAMI);
    return response.data;
  } catch (error: Error | any) {
    throw new Error(error.response?.data?.message
      || error.message || 'Whoami failed');
  }
}


export const updateProfile = async (profileData: any) => {
  try {
    const response = await axios.put(
      API.AUTH.UPDATE_PROFILE,
      profileData,
      {
        headers: {
          'Content-Type': 'multipart/form-data', // for file upload/multer
        }
      }
    );
    return response.data;
  } catch (error: Error | any) {
    throw new Error(error.response?.data?.message
      || error.message || 'Update profile failed');
  }
}
